flask_ml_server/requirements.txt
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
google-auth==2.22.0
google-auth-oauthlib==1.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.13.1
MarkupSafe==2.1.3
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
paho-mqtt==1.6.1
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-engineio==4.6.1
python-socketio==5.8.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.7
